refactor(knockdog): rename dogschool -> dog-school
