+++ apps/knockdog/src/entities/dog-school/config/constant.ts
