+++ apps/knockdog/src/entities/dog-school/index.ts
