+++ apps/knockdog/src/features/dog-school/ui/CardClipDefs.tsx
