+++ apps/knockdog/src/features/dog-school/ui/DogSchoolImageSlider.tsx
